+++ src/config/paramsSchemas/fal/flux-kontext-dev.ts
@@ -1,4 +1,4 @@
-import { type ModelParamsSchema } from '@/libs/standard-parameters/meta-schema';
+import { type ModelParamsSchema } from '@/libs/standard-parameters/index';
 
 export const fluxKontextDevParamsSchema: ModelParamsSchema = {
   imageUrl: { default: null },
