+++ src/store/image/slices/generationConfig/hooks.test.ts
@@ -2,7 +2,7 @@ import { act, renderHook } from '@testing-library/react';
 import { describe, expect, it, vi } from 'vitest';
 
 import { fluxSchnellParamsSchema } from '@/config/paramsSchemas/fal/flux-schnell';
-import { ModelParamsSchema, RuntimeImageGenParams } from '@/libs/standard-parameters/meta-schema';
+import { ModelParamsSchema, RuntimeImageGenParams } from '@/libs/standard-parameters/index';
 import { useImageStore } from '@/store/image';
 import { AIImageModelCard } from '@/types/aiModel';
