+++ src/store/image/slices/generationConfig/hooks.ts
@@ -1,10 +1,7 @@
 import { useCallback, useMemo } from 'react';
 
 import { DEFAULT_ASPECT_RATIO, PRESET_ASPECT_RATIOS } from '@/const/image';
-import {
-  RuntimeImageGenParams,
-  RuntimeImageGenParamsKeys,
-} from '@/libs/standard-parameters/meta-schema';
+import { RuntimeImageGenParams, RuntimeImageGenParamsKeys } from '@/libs/standard-parameters/index';
 
 import { useImageStore } from '../../store';
 import { imageGenerationConfigSelectors } from './selectors';
