+++ src/store/image/slices/generationConfig/initialState.ts
@@ -5,7 +5,7 @@ import {
   ModelParamsSchema,
   RuntimeImageGenParams,
   extractDefaultValues,
-} from '@/libs/standard-parameters/meta-schema';
+} from '@/libs/standard-parameters/index';
 
 export const DEFAULT_AI_IMAGE_PROVIDER = ModelProvider.OpenAI;
 export const DEFAULT_AI_IMAGE_MODEL = 'gpt-image-1';
