+++ src/store/image/slices/generationConfig/selectors.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, vi } from 'vitest';
 
 import { gptImage1ParamsSchema } from '@/config/paramsSchemas/openai/gpt-image-1';
-import { ModelParamsSchema, RuntimeImageGenParams } from '@/libs/standard-parameters/meta-schema';
+import { ModelParamsSchema, RuntimeImageGenParams } from '@/libs/standard-parameters/index';
 import { ImageStore } from '@/store/image';
 import { initialState } from '@/store/image/initialState';
 import { AIImageModelCard } from '@/types/aiModel';
